agent-service/requirements.txt
# Requirements for Agent Service

# Web framework
fastapi>=0.110.0
uvicorn>=0.31.1
python-multipart>=0.0.22

# Langchain
langchain>=1.2.13
langchain-openai>=1.1.2

# Database
sqlalchemy>=2.0.29
alembic>=1.13.1
aiosqlite>=0.20.0

# Authentication
python-jose[cryptography]>=3.5.0
passlib[bcrypt]>=1.7.4

# Environment variables
python-dotenv>=1.0.1
pydantic-settings>=2.5.2

# Logging
loguru>=0.7.2

# Utilities
pydantic>=2.11.0

# Web scraping (for web_fetch and web_search tools)
httpx>=0.28.1              # Async HTTP client (for web_fetch)
ddgs>=9.13.0               # DuckDuckGo search (for web_search)
requests>=2.31.0           # HTTP library (still needed by some dependencies)
beautifulsoup4>=4.12.3     # HTML parsing
lxml>=5.3.0                # XML/HTML parser

# Observability
langfuse>=4.0.0            # LLM tracing & observability

# LangGraph checkpoint persistence
langgraph-checkpoint-sqlite>=2.0.0

# Document processing (for read_document tool)
PyMuPDF>=1.24.0            # PDF reading
python-docx>=1.1.0         # Word reading
python-pptx>=0.6.23        # PowerPoint reading
